Python implementation of the monaural SAI.
This is a nearly-line-by-line translation of the C++ implementation and unit tests.

It will be primarily be useful for experimenting with e.g. different
triggering strategies, which is much easier to do in an interactive
Python REPL, than in C++.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=59813065