Add stabilizing log offset to pitchogram and tune default parameters.
When using log lag warping, a problem is that the plot is blurry and uses excessive space displaying the shorter lags. This commit adds a `log_offset_s` parameter to add a stabilizing offset to the lag before taking log, which limits this effect.

Additionally, this commit revises the pitchogram default parameters, including half the hop step size for finer time resolution and bilinear sampling (instead of using nearest sampling) when scaling the pitchogram to the canvas.

PiperOrigin-RevId: 462440198