Implement two_cap_with_syn in JAX, modelled to behave identically as per NumPy and MATLAB versions.
This introduces and implements a separate set of Params, Hypers, Weights and State based on a single inclusion and 'decision' point in the params, to use 'two_cap_with_syn'

Update the JAX unit tests to test the new path in addition to one_cap and two_cap.

Update the JAX carfac_float64_test to modify all ears instead of just ear 0 when perturbing.

PiperOrigin-RevId: 689152241