Update Python dependencies to allow versions preinstalled in Colab
Before this commit, installing this project in Colab was difficult:

- Running the `pip install` command in the README does not work, as
  Colab uses Python 3.11 which does not match
  `project.requires-python = ">=3.12"` in our `pyproject.toml`.
  This can be worked around by using `--ignore-requires-python`.
- The dependency versions of the libraries we use do not match the
  versions preinstalled on Colab, so installing this project changes
  those versions… breaking other dependencies such as `numba` and
  `tensorflow` / `jax` if they weren't installed as an optional
  dependency.

Updating the version specifiers in `pyproject.toml` to exactly match
what Colab fixes all of these issues, while still passing all of the
tests. However, an exact version specifier is too strict, as we know
that this code still works on the previous version numbers.

Update the version specifiers to be a **lower bound** set to the
minimum of the old version and the Colab version. These version
specifiers do not have an upper bound, but should work for the
foreseeable future. If a new version causes issues, we can add an
upper bound when needed.

Note that the Python version specified in `.python-version` was
intentionally not updated from 3.12 to 3.11, as almost all Python
features should be identical between the two.

PiperOrigin-RevId: 776486825