Format all Python files with pyink
Long strings that were split across multiple lines were unsplit before
running pyink so they can be reformatted with a consistent "space in
next line" style.

The docstring for `carfac.testing.assert_matlab_compatibility` was
manually specified to not be auto-formatted, as the formatting inserts
a line break between the dimension specification in `fun`.

PiperOrigin-RevId: 780389909