jax: Simplify `ear_ch_scan`
Taking advantage of the fact that `jax.lax.scan` can take in pytrees,
we can pass `(g, zy)` as the scanned value to get `g` and `zy` values
at each step of the scan - without needing manual indexing.

This also fixes the typing issues as `x_in` was incorrectly annotated
as a `float`, when it should have been a `jax.Array`.

PiperOrigin-RevId: 785518665