jax: Document `run_segment_jit` cache behaviour
It is non-obvious that repeatedly calling `run_segment_jit` with
different hypers or input shapes can cause memory leaks. Explicitly
document this, and suggest compiling `run_segment` locally to avoid
this.

This is most notable when running the benchmarks, which repeatedly
call `run_segment_jit` with "different" hypers and input lengths.

PiperOrigin-RevId: 788236814