Generalize non_decimating to simpler_decimating
The parallel updates in AGC filters are generalized from the non-decimating case to the "simpler-decimating" case, in which all AGC stages run at the same decimated rate.  This allows a good reduction in computation rate from the non-decimating case, with not so much complexity as the classic case.  The classic way with resursive calls to a stage update function is still supported.